requirements.txt
# pip install -r requirements.txt

mysql-connector-python
flask_mysqldb
requests